braille-reader/requirements.txt
Pillow==9.3.0
matplotlib==3.5.1
numpy==1.21.2
opencv-python==4.5.3.20210927
PyPDF2==1.26.0
python-docx==0.8.11
braille==0.1.0